requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
MarkupSafe==2.1.5
serverless-wsgi==3.0.4
sqlparse==0.5.0
Werkzeug==3.0.3
